change type

+++ incs/adjacency.h
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/13 11:13:08 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/02/13 11:14:07 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/02/19 10:42:32 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 
 t_adjacency *new_adjacency(t_room **rooms, t_graph *graph);
 
-void free_adjacency(t_adjacency *adj, int size);
+void free_adjacency(t_adjacency *adj, size_t size);
 
 #endif
