+++ srcs/adjacency/free_adjacency.c
@@ -6,16 +6,16 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/13 11:10:17 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/02/13 11:19:38 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/02/19 10:42:12 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "types.h"
 
-void free_adjacency(t_adjacency *adj, int size)
+void free_adjacency(t_adjacency *adj, size_t size)
 {
-    int index;
+    size_t index;
 
     if (adj == NULL)
         return ;
